refactor: remove impl TypePrinter for TypeSystemDocument
